feat: fix conversion of spark project to substrait (#76)
The SparkConnect project relation does not pass any fields through
except the fields
it generates. The previous behavior mimicked that of Substrait which was
to pass
all of the input fields through as well. The SparkConnect behavior is
now used.

This PR also now emits all plans emitted in failed tests, avoids
duplicate names
in joins (usually the join keys), and retains the name of projected
field references.